+++ src/constants/index.ts
@@ -1,2 +1,6 @@
 export * from "./theme";
 export * from "./home";
+export * from "./nav";
+export * from "./copy";
+export * from "./packages";
+export * from "./about";